@@ -29,7 +29,7 @@ class _PhoneVerificationState extends State<PhoneVerificationView> {
   ];
 
   int currentIndex = 0;
-  String errorMsg;
+  String? errorMsg;
   bool loading = false;
 
   @override
@@ -174,7 +174,7 @@ class _PhoneVerificationState extends State<PhoneVerificationView> {
             Padding(
               padding: EdgeInsets.only(bottom: 8),
               child: Text(
-                errorMsg,
+                errorMsg!,
                 textAlign: TextAlign.center,
                 style: TextStyle(
                   color: SKColors.warning_red,
