@@ -16,18 +16,17 @@ class JobsView extends StatefulWidget {
 enum _ProfileState { intro, start, resume, profile }
 
 class _JobsViewState extends State<JobsView> {
-  _ProfileState profileState;
-
-  TypeObject jobType;
-  DateTime graduationDate;
+  _ProfileState? profileState;
+  TypeObject? jobType;
+  DateTime? graduationDate;
 
   @override
   void initState() {
     super.initState();
 
-    if (SKUser.current.student.gradYear != null)
+    if (SKUser.current!.student.gradYear != null)
       graduationDate =
-          DateTime.parse('${SKUser.current.student.gradYear}-05-01');
+          DateTime.parse('${SKUser.current!.student.gradYear}-05-01');
 
     updateProfileState();
 
@@ -36,7 +35,7 @@ class _JobsViewState extends State<JobsView> {
       channel: NotificationChannels.newTabSelected,
       onNotification: (index) {
         if (index == JOBS_TAB)
-          SKUser.current.getJobProfile().then((response) {
+          SKUser.current!.getJobProfile().then((response) {
             if (response.wasSuccessful()) {
               updateProfileState();
               setState(() {});
@@ -81,7 +80,7 @@ class _JobsViewState extends State<JobsView> {
     final result = await showDialog(
       context: context,
       builder: (_) => _GraduationDatePicker(
-        startDate: graduationDate,
+        startDate: graduationDate!,
       ),
     );
 
@@ -123,7 +122,7 @@ class _JobsViewState extends State<JobsView> {
           items: items.map((t) => t.name).toList(),
           onSelect: (index) async {
             final loader = SKLoadingScreen.fadeIn(context);
-            await SKUser.current.update(degreeType: items[index]);
+            await SKUser.current!.update(degreeType: items[index]);
             loader.fadeOut();
 
             setState(() {});
@@ -170,13 +169,13 @@ class _JobsViewState extends State<JobsView> {
 
     if (JobProfile.currentProfile == null) {
       response = await JobProfile.createProfile(
-        jobType: jobType,
-        graduationDate: graduationDate,
+        jobType: jobType!,
+        graduationDate: graduationDate!,
       );
     } else {
-      response = await JobProfile.currentProfile.updateProfile(
-        jobSearchType: jobType,
-        gradDate: graduationDate,
+      response = await JobProfile.currentProfile!.updateProfile(
+        jobSearchType: jobType!,
+        gradDate: graduationDate!,
       );
     }
 
@@ -184,7 +183,7 @@ class _JobsViewState extends State<JobsView> {
 
     if (response.wasSuccessful()) {
       setState(() {
-        profileState = JobProfile.currentProfile.resume_url == null
+        profileState = JobProfile.currentProfile!.resume_url == null
             ? _ProfileState.resume
             : _ProfileState.profile;
       });
@@ -202,7 +201,7 @@ class _JobsViewState extends State<JobsView> {
 
   @override
   Widget build(BuildContext context) {
-    List<Widget> children;
+   late  List<Widget> children;
 
     switch (profileState) {
       case _ProfileState.intro:
@@ -290,7 +289,7 @@ class _JobsViewState extends State<JobsView> {
       ];
 
   List<Widget> createStart() {
-    final fields = SKUser.current.student.fieldsOfStudy ?? <FieldsOfStudy>[];
+    final fields = SKUser.current!.student.fieldsOfStudy ?? <FieldsOfStudy>[];
     String fieldsBody;
 
     if (fields.length > 0)
@@ -300,7 +299,7 @@ class _JobsViewState extends State<JobsView> {
 
     final isValid = graduationDate != null &&
         fields.length > 0 &&
-        SKUser.current.student.degreeType != null &&
+        SKUser.current!.student.degreeType != null &&
         jobType != null;
 
     return [
@@ -351,7 +350,7 @@ class _JobsViewState extends State<JobsView> {
                       Text(
                         graduationDate == null
                             ? 'Select...'
-                            : DateFormat('MMMM yyyy').format(graduationDate),
+                            : DateFormat('MMMM yyyy').format(graduationDate!),
                         style: TextStyle(
                             color: graduationDate == null
                                 ? SKColors.jobs_light_green
@@ -424,10 +423,10 @@ class _JobsViewState extends State<JobsView> {
                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
                     children: [
                       Text(
-                        SKUser.current.student.degreeType?.name ?? 'Select...',
+                        SKUser.current!.student.degreeType?.name ?? 'Select...',
                         style: TextStyle(
                             color:
-                                SKUser.current.student.degreeType?.name == null
+                                SKUser.current!.student.degreeType?.name == null
                                     ? SKColors.jobs_light_green
                                     : SKColors.jobs_dark_green),
                       ),
@@ -572,7 +571,7 @@ class _JobsViewState extends State<JobsView> {
               child: Container(
                 alignment: Alignment.center,
                 child: _SKJobProfileCompletionCircle(
-                  completion: JobProfile.currentProfile.profile_score,
+                  completion: JobProfile.currentProfile!.profile_score,
                 ),
               ),
             ),
@@ -620,7 +619,7 @@ class _JobsViewState extends State<JobsView> {
     String type;
     bool startsWithVowel;
 
-    switch (JobProfile.currentProfile.job_search_type.id) {
+    switch (JobProfile.currentProfile!.job_search_type.id) {
       case 100:
         type = ' internship';
         startsWithVowel = true;
@@ -647,7 +646,7 @@ class _JobsViewState extends State<JobsView> {
 }
 
 class _GraduationDatePicker extends StatefulWidget {
-  final DateTime startDate;
+  final DateTime? startDate;
 
   _GraduationDatePicker({this.startDate});
 
@@ -656,14 +655,14 @@ class _GraduationDatePicker extends StatefulWidget {
 }
 
 class _GraduationDatePickerState extends State<_GraduationDatePicker> {
-  int monthIndex;
-  String year;
+  int? monthIndex;
+  String? year;
 
-  List<String> months;
-  List<String> years;
+  List<String>? months;
+  List<String>? years;
 
-  ScrollController monthController;
-  ScrollController yearController;
+  var monthController;
+  var yearController;
 
   @override
   void initState() {
@@ -689,22 +688,22 @@ class _GraduationDatePickerState extends State<_GraduationDatePicker> {
     ];
 
     if (widget.startDate != null) {
-      monthIndex = widget.startDate.month - 1;
-      year = '${widget.startDate.year}';
+      monthIndex = widget.startDate!.month - 1;
+      year = '${widget.startDate!.year}';
     } else {
       monthIndex = 0;
-      year = years.first;
+      year = years!.first;
     }
 
-    monthController = FixedExtentScrollController(initialItem: monthIndex);
+    monthController = FixedExtentScrollController(initialItem: monthIndex!);
     yearController =
-        FixedExtentScrollController(initialItem: years.indexOf(year));
+        FixedExtentScrollController(initialItem: years!.indexOf(year!));
   }
 
   @override
   void dispose() {
-    monthController.dispose();
-    yearController.dispose();
+    monthController!.dispose();
+    yearController!.dispose();
 
     super.dispose();
   }
@@ -748,10 +747,10 @@ class _GraduationDatePickerState extends State<_GraduationDatePicker> {
                         itemBuilder: (_, index) => Container(
                           alignment: Alignment.center,
                           child: Text(
-                            months[index],
+                            months![index],
                             style: Theme.of(context)
                                 .textTheme
-                                .body1
+                                .body1!
                                 .copyWith(fontWeight: FontWeight.normal),
                           ),
                         ),
@@ -764,15 +763,15 @@ class _GraduationDatePickerState extends State<_GraduationDatePicker> {
                         itemExtent: 24,
                         scrollController: yearController,
                         onSelectedItemChanged: (yearIndex) =>
-                            this.year = years[yearIndex],
-                        childCount: years.length,
+                            this.year = years![yearIndex],
+                        childCount: years!.length,
                         itemBuilder: (_, index) => Container(
                           alignment: Alignment.center,
                           child: Text(
-                            years[index],
+                            years![index],
                             style: Theme.of(context)
                                 .textTheme
-                                .body1
+                                .body1!
                                 .copyWith(fontWeight: FontWeight.normal),
                           ),
                         ),
@@ -814,7 +813,7 @@ class _GraduationDatePickerState extends State<_GraduationDatePicker> {
                     child: GestureDetector(
                       behavior: HitTestBehavior.opaque,
                       onTapUp: (_) {
-                        final monthNum = monthIndex + 1;
+                        final monthNum = monthIndex! + 1;
                         String monthStr = '$monthNum';
                         if (monthNum < 10) monthStr = '0$monthNum';
 
@@ -845,7 +844,7 @@ class _GraduationDatePickerState extends State<_GraduationDatePicker> {
 }
 
 class _SKJobProfileCompletionCircle extends StatefulWidget {
-  final num completion;
+  final num? completion;
 
   _SKJobProfileCompletionCircle({this.completion});
 
@@ -856,8 +855,8 @@ class _SKJobProfileCompletionCircle extends StatefulWidget {
 class _SKJobProfileCompletionCircleState
     extends State<_SKJobProfileCompletionCircle>
     with SingleTickerProviderStateMixin {
-  AnimationController animationController;
-  Animation<double> animation;
+  AnimationController? animationController;
+  late Animation<double> animation;
 
   @override
   void initState() {
@@ -869,45 +868,45 @@ class _SKJobProfileCompletionCircleState
     );
     animation = createAnimation();
 
-    animationController.forward();
+    animationController!.forward();
 
     DartNotificationCenter.subscribe(
         observer: this,
         channel: NotificationChannels.newTabSelected,
         onNotification: (index) {
-          if (index == JOBS_TAB && !animationController.isAnimating) {
-            animationController.forward(from: 0);
+          if (index == JOBS_TAB && !animationController!.isAnimating) {
+            animationController!.forward(from: 0);
           }
         });
   }
 
   @override
   void didUpdateWidget(Widget oldWidget) {
-    super.didUpdateWidget(oldWidget);
+    //super.didUpdateWidget(oldWidget);
     if ((oldWidget as _SKJobProfileCompletionCircle).completion !=
         widget.completion) {
       animation.removeListener(animationListener);
 
       animation = createAnimation();
-      if (!animationController.isAnimating)
-        animationController.forward(from: 0);
+      if (!animationController!.isAnimating)
+        animationController!.forward(from: 0);
     }
   }
 
   @override
   void dispose() {
     DartNotificationCenter.unsubscribe(observer: this);
-    animationController.dispose();
+    animationController!.dispose();
 
     super.dispose();
   }
 
   void animationListener() => setState(() {});
 
-  Animation createAnimation() =>
-      Tween<double>(begin: 0, end: widget.completion).animate(
+  Animation<double> createAnimation() =>
+      Tween<double>(begin: 0, end: widget.completion?.toDouble()).animate(
         CurvedAnimation(
-            parent: animationController, curve: Curves.easeOutCubic),
+            parent: animationController!, curve: Curves.easeOutCubic),
       )..addListener(animationListener);
 
   @override
