@@ -14,24 +14,24 @@ class ClassSearchSettingsModal extends StatefulWidget {
 }
 
 class _ClassSearchSettingsModalState extends State<ClassSearchSettingsModal> {
-  School school;
-  Period period;
+  School? school;
+  Period? period;
 
   List<Period> possiblePeriods = [];
 
   @override
   void initState() {
-    school = SKUser.current.student.primarySchool;
+    school = SKUser.current!.student.primarySchool;
 
     if (widget.initialPeriodId != null) {
       period = Period.currentPeriods[widget.initialPeriodId];
 
-      if (period.id != SKUser.current.student.primaryPeriod.id)
-        SKUser.current.update(primaryPeriod: period);
+      if (period!.id != SKUser.current!.student.primaryPeriod?.id)
+        SKUser.current!.update(primaryPeriod: period!);
     }
 
     final now = DateTime.now();
-    possiblePeriods = (school.periods.toList())
+    possiblePeriods = (school!.periods.toList())
       ..removeWhere(
         (p) =>
             now.isAfter(p.endDate.add(Duration(days: 15))) ||
@@ -59,7 +59,7 @@ class _ClassSearchSettingsModalState extends State<ClassSearchSettingsModal> {
       period = null;
 
       final now = DateTime.now();
-      possiblePeriods = (school.periods.toList())
+      possiblePeriods = (school!.periods.toList())
         ..removeWhere(
           (p) =>
               now.isAfter(p.endDate.add(Duration(days: 15))) ||
@@ -72,7 +72,7 @@ class _ClassSearchSettingsModalState extends State<ClassSearchSettingsModal> {
     }
   }
 
-  void tappedPeriod(TapUpDetails details) async {
+  void tappedPeriod(TapUpDetails? details) async {
     showDialog(
       context: context,
       builder: (context) => SKPickerModal(
@@ -82,7 +82,7 @@ class _ClassSearchSettingsModalState extends State<ClassSearchSettingsModal> {
           setState(() {
             period = possiblePeriods[index];
           });
-          SKUser.current.update(primaryPeriod: period);
+          SKUser.current!.update(primaryPeriod: period!);
         },
       ),
     );
