@@ -11,15 +11,15 @@ class TodoPreferencesModal extends StatefulWidget {
 }
 
 class _TodoPreferencesState extends State<TodoPreferencesModal> {
-  int daysPast;
-  int daysFuture;
+  late int daysPast;
+  late int daysFuture;
 
   @override
   void initState() {
     super.initState();
 
-    daysFuture = SKUser.current.student.todoDaysFuture;
-    daysPast = SKUser.current.student.todoDaysPast;
+    daysFuture = SKUser.current!.student.todoDaysFuture!;
+    daysPast = SKUser.current!.student.todoDaysPast!;
   }
 
   void tappedNotifications(_) => DartNotificationCenter.post(
@@ -65,7 +65,7 @@ class _TodoPreferencesState extends State<TodoPreferencesModal> {
   }
 
   void tappedSave(_) async {
-    final student = SKUser.current.student;
+    final student = SKUser.current!.student;
 
     if (daysFuture == student.todoDaysFuture &&
         daysPast == student.todoDaysPast)
@@ -73,7 +73,7 @@ class _TodoPreferencesState extends State<TodoPreferencesModal> {
     else {
       final loader = SKLoadingScreen.fadeIn(context);
 
-      final result = await SKUser.current
+      final result = await SKUser.current!
           .update(todoDaysFuture: daysFuture, todoDaysPast: daysPast);
 
       loader.fadeOut();
@@ -91,7 +91,7 @@ class _TodoPreferencesState extends State<TodoPreferencesModal> {
 
   @override
   Widget build(BuildContext context) {
-    final student = SKUser.current.student;
+    final student = SKUser.current!.student;
     final shouldSave = daysPast != student.todoDaysPast ||
         daysFuture != student.todoDaysFuture;
 
